+++ debezium-core/src/main/java/io/debezium/pipeline/source/snapshot/incremental/AbstractIncrementalSnapshotChangeEventSource.java
@@ -17,12 +17,10 @@ import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.OptionalLong;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -45,7 +43,6 @@ import io.debezium.pipeline.spi.ChangeRecordEmitter;
 import io.debezium.pipeline.spi.OffsetContext;
 import io.debezium.pipeline.spi.Partition;
 import io.debezium.relational.Column;
-import io.debezium.relational.Key.KeyMapper;
 import io.debezium.relational.RelationalDatabaseConnectorConfig;
 import io.debezium.relational.RelationalDatabaseSchema;
 import io.debezium.relational.RelationalSnapshotChangeEventSource;
@@ -53,7 +50,6 @@ import io.debezium.relational.SnapshotChangeRecordEmitter;
 import io.debezium.relational.Table;
 import io.debezium.relational.TableId;
 import io.debezium.relational.TableSchema;
-import io.debezium.relational.Tables.ColumnNameFilter;
 import io.debezium.schema.DatabaseSchema;
 import io.debezium.spi.schema.DataCollectionId;
 import io.debezium.util.Clock;
@@ -83,7 +79,7 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
     protected EventDispatcher<P, T> dispatcher;
     protected IncrementalSnapshotContext<T> context = null;
     protected JdbcConnection jdbcConnection;
-    protected ColumnNameFilter columnFilter;
+    protected ChunkQueryBuilder<T> chunkQueryBuilder;
     protected final Map<Struct, Object[]> window = new LinkedHashMap<>();
     protected final NotificationService<P, ? extends OffsetContext> notificationService;
 
@@ -97,7 +93,7 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
                                                         NotificationService<P, ? extends OffsetContext> notificationService) {
         this.connectorConfig = config;
         this.jdbcConnection = jdbcConnection;
-        this.columnFilter = config.getColumnFilter();
+        this.chunkQueryBuilder = jdbcConnection.chunkQueryBuilder(config);
         this.dispatcher = dispatcher;
         this.databaseSchema = (RelationalDatabaseSchema) databaseSchema;
         this.clock = clock;
@@ -215,88 +211,6 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
      */
     protected abstract void emitWindowClose(P partition, OffsetContext offsetContext) throws Exception;
 
-    protected String buildChunkQuery(Table table, Optional<String> additionalCondition) {
-        return buildChunkQuery(table, connectorConfig.getIncrementalSnapshotChunkSize(), additionalCondition);
-    }
-
-    protected String buildChunkQuery(Table table, int limit, Optional<String> additionalCondition) {
-        String condition = null;
-        // Add condition when this is not the first query
-        if (context.isNonInitialChunk()) {
-            final StringBuilder sql = new StringBuilder();
-            // Window boundaries
-            addLowerBound(table, sql);
-            // Table boundaries
-            sql.append(" AND NOT ");
-            addLowerBound(table, sql);
-            condition = sql.toString();
-        }
-        final String orderBy = getQueryColumns(table).stream()
-                .map(c -> jdbcConnection.quotedColumnIdString(c.name()))
-                .collect(Collectors.joining(", "));
-        return jdbcConnection.buildSelectWithRowLimits(table.id(),
-                limit,
-                buildProjection(table),
-                Optional.ofNullable(condition),
-                additionalCondition,
-                orderBy);
-    }
-
-    protected String buildProjection(Table table) {
-        String projection = "*";
-        if (connectorConfig.isColumnsFiltered()) {
-            TableId tableId = table.id();
-            projection = table.columns().stream()
-                    .filter(column -> columnFilter.matches(tableId.catalog(), tableId.schema(), tableId.table(), column.name()))
-                    .map(column -> jdbcConnection.quotedColumnIdString(column.name()))
-                    .collect(Collectors.joining(", "));
-        }
-        return projection;
-    }
-
-    private void addLowerBound(Table table, StringBuilder sql) {
-        // To make window boundaries working for more than one column it is necessary to calculate
-        // with independently increasing values in each column independently.
-        // For one column the condition will be (? will always be the last value seen for the given column)
-        // (k1 > ?)
-        // For two columns
-        // (k1 > ?) OR (k1 = ? AND k2 > ?)
-        // For four columns
-        // (k1 > ?) OR (k1 = ? AND k2 > ?) OR (k1 = ? AND k2 = ? AND k3 > ?) OR (k1 = ? AND k2 = ? AND k3 = ? AND k4 > ?)
-        // etc.
-        final List<Column> pkColumns = getQueryColumns(table);
-        if (pkColumns.size() > 1) {
-            sql.append('(');
-        }
-        for (int i = 0; i < pkColumns.size(); i++) {
-            final boolean isLastIterationForI = (i == pkColumns.size() - 1);
-            sql.append('(');
-            for (int j = 0; j < i + 1; j++) {
-                final boolean isLastIterationForJ = (i == j);
-                sql.append(jdbcConnection.quotedColumnIdString(pkColumns.get(j).name()));
-                sql.append(isLastIterationForJ ? " > ?" : " = ?");
-                if (!isLastIterationForJ) {
-                    sql.append(" AND ");
-                }
-            }
-            sql.append(")");
-            if (!isLastIterationForI) {
-                sql.append(" OR ");
-            }
-        }
-        if (pkColumns.size() > 1) {
-            sql.append(')');
-        }
-    }
-
-    protected String buildMaxPrimaryKeyQuery(Table table, Optional<String> additionalCondition) {
-        final String orderBy = getQueryColumns(table).stream()
-                .map(c -> jdbcConnection.quotedColumnIdString(c.name()))
-                .collect(Collectors.joining(" DESC, ")) + " DESC";
-        return jdbcConnection.buildSelectWithRowLimits(table.id(), 1, buildProjection(table), Optional.empty(),
-                additionalCondition, orderBy);
-    }
-
     @Override
     @SuppressWarnings("unchecked")
     public void init(P partition, OffsetContext offsetContext) {
@@ -359,7 +273,7 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
                     Object[] maximumKey;
                     try {
                         maximumKey = jdbcConnection.queryAndMap(
-                                buildMaxPrimaryKeyQuery(currentTable, context.currentDataCollectionId().getAdditionalCondition()), rs -> {
+                                chunkQueryBuilder.buildMaxPrimaryKeyQuery(context, currentTable, context.currentDataCollectionId().getAdditionalCondition()), rs -> {
                                     if (!rs.next()) {
                                         return null;
                                     }
@@ -439,8 +353,8 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
             nextDataCollection(partition, offsetContext);
             return true;
         }
-        if (getQueryColumns(currentTable).isEmpty()) {
-            LOGGER.warn("Incremental snapshot for table '{}' skipped cause the table has no primary keys", currentTableId);
+        if (chunkQueryBuilder.getQueryColumns(context, currentTable).isEmpty()) {
+            LOGGER.warn("Incremental snapshot for table '{}' skipped because the table has no primary keys", currentTableId);
             notificationService.incrementalSnapshotNotificationService().notifyTableScanCompleted(context, partition, offsetContext, totalRowsScanned, NO_PRIMARY_KEY);
             nextDataCollection(partition, offsetContext);
             return true;
@@ -480,10 +394,10 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
     }
 
     private Table readSchema() {
-        final String selectStatement = buildChunkQuery(currentTable, 0, Optional.empty());
+        final String selectStatement = chunkQueryBuilder.buildChunkQuery(context, currentTable, 0, Optional.empty());
         LOGGER.debug("Reading schema for table '{}' using select statement: '{}'", currentTable.id(), selectStatement);
 
-        try (PreparedStatement statement = readTableChunkStatement(selectStatement);
+        try (PreparedStatement statement = chunkQueryBuilder.readTableChunkStatement(context, currentTable, selectStatement);
                 ResultSet rs = statement.executeQuery()) {
             return getTable(rs);
         }
@@ -608,16 +522,6 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
         }
     }
 
-    protected void addKeyColumnsToCondition(Table table, StringBuilder sql, String predicate) {
-        for (Iterator<Column> i = getQueryColumns(table).iterator(); i.hasNext();) {
-            final Column key = i.next();
-            sql.append(jdbcConnection.quotedColumnIdString(key.name())).append(predicate);
-            if (i.hasNext()) {
-                sql.append(" AND ");
-            }
-        }
-    }
-
     /**
      * Expands the string-based list of data collection ids if supplied using regex to a list of
      * all matching explicit data collection ids.
@@ -644,13 +548,13 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
         long exportStart = clock.currentTimeInMillis();
         LOGGER.debug("Exporting data chunk from table '{}' (total {} tables)", currentTable.id(), context.dataCollectionsToBeSnapshottedCount());
 
-        final String selectStatement = buildChunkQuery(currentTable, context.currentDataCollectionId().getAdditionalCondition());
+        final String selectStatement = chunkQueryBuilder.buildChunkQuery(context, currentTable, context.currentDataCollectionId().getAdditionalCondition());
         LOGGER.debug("\t For table '{}' using select statement: '{}', key: '{}', maximum key: '{}'", currentTable.id(),
                 selectStatement, context.chunkEndPosititon(), context.maximumKey().get());
 
         final TableSchema tableSchema = databaseSchema.schemaFor(currentTable.id());
 
-        try (PreparedStatement statement = readTableChunkStatement(selectStatement);
+        try (PreparedStatement statement = chunkQueryBuilder.readTableChunkStatement(context, currentTable, selectStatement);
                 ResultSet rs = statement.executeQuery()) {
             if (checkSchemaChanges(rs)) {
                 return false;
@@ -748,30 +652,6 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
         progressListener.currentChunk(partition, null, null, null, null);
     }
 
-    protected PreparedStatement readTableChunkStatement(String sql) throws SQLException {
-        final PreparedStatement statement = jdbcConnection.readTablePreparedStatement(connectorConfig, sql,
-                OptionalLong.empty());
-        if (context.isNonInitialChunk()) {
-            final Object[] maximumKey = context.maximumKey().get();
-            final Object[] chunkEndPosition = context.chunkEndPosititon();
-            // Fill boundaries placeholders
-            int pos = 0;
-            final List<Column> queryColumns = getQueryColumns(currentTable);
-            for (int i = 0; i < chunkEndPosition.length; i++) {
-                for (int j = 0; j < i + 1; j++) {
-                    jdbcConnection.setQueryColumnValue(statement, queryColumns.get(j), ++pos, chunkEndPosition[j]);
-                }
-            }
-            // Fill maximum key placeholders
-            for (int i = 0; i < chunkEndPosition.length; i++) {
-                for (int j = 0; j < i + 1; j++) {
-                    jdbcConnection.setQueryColumnValue(statement, queryColumns.get(j), ++pos, maximumKey[j]);
-                }
-            }
-        }
-        return statement;
-    }
-
     private Timer getTableScanLogTimer() {
         return Threads.timer(clock, RelationalSnapshotChangeEventSource.LOG_INTERVAL);
     }
@@ -781,7 +661,7 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
         if (row == null) {
             return null;
         }
-        final List<Column> keyColumns = getQueryColumns(currentTable);
+        final List<Column> keyColumns = chunkQueryBuilder.getQueryColumns(context, currentTable);
         final Object[] key = new Object[keyColumns.size()];
         for (int i = 0; i < keyColumns.size(); i++) {
             final Object fieldValue = row[keyColumns.get(i).position() - 1];
@@ -822,19 +702,5 @@ public abstract class AbstractIncrementalSnapshotChangeEventSource<P extends Par
         // since schema changes are not emitted as change events in the same way that they are for
         // connectors like MySQL or Oracle
         return table;
-    }
-
-    private KeyMapper getKeyMapper() {
-        return connectorConfig.getKeyMapper() == null ? table -> table.primaryKeyColumns() : connectorConfig.getKeyMapper();
-    }
-
-    private List<Column> getQueryColumns(Table table) {
-        if (context != null && context.currentDataCollectionId() != null) {
-            Optional<String> surrogateKey = context.currentDataCollectionId().getSurrogateKey();
-            if (surrogateKey.isPresent()) {
-                return Collections.singletonList(table.columnWithName(surrogateKey.get()));
-            }
-        }
-        return getKeyMapper().getKeyKolumns(table);
     }
 }
